backend/requirements.txt
alembic==1.4.3
Authlib==0.14.3
fastapi==0.63.0
celery==5.0.0
redis==3.5.3
httpx==0.15.5
ipython==7.18.1
itsdangerous==1.1.0
Jinja2==2.11.2
psycopg2-binary==2.8.6
pytest==6.1.0
requests==2.24.0
SQLAlchemy==1.3.19
uvicorn==0.12.1
passlib==1.7.2
bcrypt==3.2.0
sqlalchemy-utils==0.36.8
python-multipart==0.0.5
pyjwt==1.7.1
scikit-learn==0.24.0
ffmpeg-python==0.2.0
tqdm==4.55.1
webvtt-py==0.4.6
beautifulsoup4==4.9.3
spacy==2.3.5